Replace black with ruff (#18)
black depends on pathspec that is licensed under MPL 2.0. Even though
MPL 2.0 can _maybe_ be included in an ASF project, I'd
prefer to simplify the discussion and just replace black with ruff that
is licensed as Apache 2.0, does not have any dependencies, and is
already used by other Apache projects.